Rename classes and separate work into distinct visitors
